pyproject.toml
[project]
name = "discord-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord.py",
    "aiosqlite",
    "numpy",
    "scipy",
    "matplotlib",
    "requests",
    "pandas",
]
